included all records on purchase
